# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octara API
  description: https://octara.xyz/docs/oauth-guide.md
  version: 1.0.0
  contact:
    name: Octara Team
    url: https://octara.xyz
servers:
  - url: https://octara.xyz/api
    description: Production environment
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://octara.xyz/oauth/authorize
          tokenUrl: https://octara.xyz/api/oauth/token
          scopes:
            read:profile: Access core profile information (display name, id)
            read:email: Access user email address
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        description:
          type: string
paths:
...
  /v1/me/avatar:
    get:
      tags: [Resource]
      summary: Get Current User Avatar
      description: Returns the avatar image of the authenticated user.
      security:
        - OAuth2: [read:profile]
      responses:
        "200":
          description: Avatar image
          content:
            image/*:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...